+++ lib/feature/force_event/presentation/overlay_force_update_dialog.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../core/common_provider/launch_url.dart';
-import '../../../core/common_widget/button/primary_filled_button.dart';
+import '../../../core/common_widget/button/filled_button.dart';
 import '../../../util/constant/url.dart';
 import '../../../util/extension/target_platform_extension.dart';
