+++ lib/feature/introduction/presentation/introduction_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../core/common_provider/dialog_controller.dart';
 import '../../../core/common_provider/launch_url.dart';
-import '../../../core/common_widget/button/primary_filled_button.dart';
+import '../../../core/common_widget/button/filled_button.dart';
 import '../../../util/constant/url.dart';
 import 'confirm_agreement_dialog.dart';
